feat: 新增 StorageManager 的清理與顯示功能測試
- 新增 storageManagerTestHarness.js，提供測試所需的輔助函數，包括 Blob polyfill、DOM 結構建立、Chrome mock 物件等。
- 新增 StorageManager.cleanup.test.js，測試 StorageManager 的 executeUnifiedCleanup 方法，確保在不同情況下的行為正確，包括無清理計劃、成功清理、清理失敗等情境。
- 新增 StorageManager.display.test.js，測試 StorageManager 的使用量與健康狀態顯示功能，確保 UI 正確更新並顯示相應的狀態與警告。
- 新增 optionsHtmlStructure.test.js，檢查 options.html 的結構與可訪問性，確保 HTML 元素符合預期的標準與設計要求。

這些變更旨在增強測試覆蓋率，確保 StorageManager 的功能在不同情況下均能正常運作，並提升使用者介面的可用性與可訪問性。
